requirements.txt
undetected-chromedriver
selenium
boto3
python-dotenv
